[dev]ホテル掲載ページの作成#5
