+++ src/components/HotelOfUserProfile.tsx
@@ -1,0 +1,12 @@
+import React from "react";
+import { HotelDetailType } from "types/types";
+
+type PROPS = {
+  props: HotelDetailType;
+};
+
+const HotelOfUserProfile = ({ props }: PROPS) => {
+  return <div></div>;
+};
+
+export default HotelOfUserProfile;
